+++ src/compas/geometry/vector.py
@@ -2,8 +2,6 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 
-from compas import PRECISION
-
 from compas.geometry import length_vector
 from compas.geometry import cross_vectors
 from compas.geometry import subtract_vectors
@@ -82,12 +80,11 @@ class Vector(Geometry):
         self.z = z
 
     def __repr__(self):
-        return "{0}(x={1:.{4}f}, y={2:.{4}f}, z={3:.{4}f})".format(
+        return "{0}(x={1}, y={2}, z={3})".format(
             type(self).__name__,
             self.x,
             self.y,
             self.z,
-            PRECISION[:1],
         )
 
     def __len__(self):